Fix type conversion for clang

+++ inst/include/dust/interface.hpp
@@ -485,7 +485,8 @@ cpp11::sexp dust_compare_data(SEXP ptr) {
   const size_t n_pars = obj->n_pars();
   if (n_pars > 0) {
     const size_t n_particles_each = obj->n_particles() / n_pars;
-    ret_r.attr("dim") = cpp11::writable::integers({n_particles_each, n_pars});
+    ret_r.attr("dim") = cpp11::writable::integers({(int)n_particles_each,
+                                                   (int)n_pars});
   }
 
   return ret_r;
